requirements/train_reqs.txt
asn1crypto==0.23.0
backports-abc==0.5
bcrypt==3.1.4
bleach==1.5.0
certifi==2017.11.5
cffi==1.11.2
configparser==3.5.0
cryptography==1.9
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
enum34==1.1.6
html5lib==0.9999999
httpsig-cffi==15.0.0
idna==2.6
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.3
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==4.1.1
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.1
numpy==1.13.3
oci==1.3.11
olefile==0.44
opencv-python==3.4.0.12
pandas==0.21.0
pandocfilters==1.4.2
paramiko==2.4.0
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.1
pyocr==0.5
pyparsing==2.2.0
pypillowfight==0.2.2
pytesseract==0.1.8
python-dateutil==2.5.3
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.11.1
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc2
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
typing==3.6.2
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.0.6
